+++ src/users/entities/entity.ts
@@ -1,4 +1,4 @@
-export interface User {
+export class User {
   id: string;
   login: string;
   password: string;
